+++ scispacy/linking.py
@@ -16,13 +16,13 @@ class EntityLinker:
     entities above a given threshold.
 
 
-    This class sets the `._.umls_ents` attribute on spacy Spans, which consists of a
-    List[Tuple[str, float]] corresponding to the UMLS concept_id and the associated score
+    This class sets the `._.kb_ents` attribute on spacy Spans, which consists of a
+    List[Tuple[str, float]] corresponding to the KB concept_id and the associated score
     for a list of `max_entities_per_mention` number of entities.
 
-    You can look up more information for a given id using the umls attribute of this class:
+    You can look up more information for a given id using the kb attribute of this class:
 
-    print(linker.umls.cui_to_entity[concept_id])
+    print(linker.kb.cui_to_entity[concept_id])
 
     A Note on Definitions:
     Only 187767 entities, or 6.74% of the UMLS KB have definitions. However,
@@ -69,17 +69,21 @@ class EntityLinker:
         max_entities_per_mention: int = 5,
         name: str = None,
     ):
-
+        # TODO(Mark): Remove in scispacy v1.0.
         Span.set_extension("umls_ents", default=[], force=True)
+        Span.set_extension("kb_ents", default=[], force=True)
 
         self.candidate_generator = candidate_generator or CandidateGenerator(name=name)
         self.resolve_abbreviations = resolve_abbreviations
         self.k = k
         self.threshold = threshold
         self.no_definition_threshold = no_definition_threshold
-        self.umls = self.candidate_generator.umls
+        self.kb = self.candidate_generator.kb
         self.filter_for_definitions = filter_for_definitions
         self.max_entities_per_mention = max_entities_per_mention
+
+        # TODO(Mark): Remove in scispacy v1.0. This is for backward compatability only.
+        self.umls = self.kb
 
     def __call__(self, doc: Doc) -> Doc:
         mentions = []
@@ -104,7 +108,7 @@ class EntityLinker:
                 score = max(cand.similarities)
                 if (
                     self.filter_for_definitions
-                    and self.umls.cui_to_entity[cand.concept_id].definition is None
+                    and self.kb.cui_to_entity[cand.concept_id].definition is None
                     and score < self.no_definition_threshold
                 ):
                     continue
@@ -112,5 +116,6 @@ class EntityLinker:
                     predicted.append((cand.concept_id, score))
             sorted_predicted = sorted(predicted, reverse=True, key=lambda x: x[1])
             mention._.umls_ents = sorted_predicted[: self.max_entities_per_mention]
+            mention._.kb_ents = sorted_predicted[: self.max_entities_per_mention]
 
         return doc
