+++ tests/test_linking.py
@@ -4,11 +4,12 @@ import tempfile
 import spacy
 
 from scispacy.candidate_generation import CandidateGenerator, create_tfidf_ann_index
-from scispacy.umls_linking import UmlsEntityLinker
+from scispacy.linking import EntityLinker
 from scispacy.umls_utils import UmlsKnowledgeBase
 from scispacy.abbreviation import AbbreviationDetector
 
-class TestUmlsLinker(unittest.TestCase):
+
+class TestLinker(unittest.TestCase):
     def setUp(self):
         super().setUp()
         self.nlp = spacy.load("en_core_web_sm")
@@ -18,7 +19,7 @@ class TestUmlsLinker(unittest.TestCase):
             umls_concept_aliases, tfidf_vectorizer, ann_index = create_tfidf_ann_index(dir_name, umls_fixture)
         candidate_generator = CandidateGenerator(ann_index, tfidf_vectorizer, umls_concept_aliases, umls_fixture)
 
-        self.linker = UmlsEntityLinker(candidate_generator, filter_for_definitions=False)
+        self.linker = EntityLinker(candidate_generator, filter_for_definitions=False)
 
     def test_naive_entity_linking(self):
         text = "There was a lot of Dipalmitoylphosphatidylcholine."
@@ -30,24 +31,24 @@ class TestUmlsLinker(unittest.TestCase):
         self.linker.filter_for_definitions = True
         self.linker.no_definition_threshold = 3.0
         doc = self.linker(doc)
-        assert doc.ents[0]._.umls_ents == []
+        assert doc.ents[0]._.kb_ents == []
 
         # Check that the linker returns only high confidence entities if we
         # set the threshold to something more reasonable.
         self.linker.no_definition_threshold = 0.95
         doc = self.linker(doc)
-        assert doc.ents[0]._.umls_ents == [('C0000039', 1.0)]
+        assert doc.ents[0]._.kb_ents == [('C0000039', 1.0)]
 
         self.linker.filter_for_definitions = False
         self.linker.threshold = 0.45
         doc = self.linker(doc)
         # Without the filter_for_definitions filter, we get 2 entities for
         # the first mention.
-        assert len(doc.ents[0]._.umls_ents) == 2
+        assert len(doc.ents[0]._.kb_ents) == 2
 
-        id_with_score = doc.ents[0]._.umls_ents[0]
+        id_with_score = doc.ents[0]._.kb_ents[0]
         assert id_with_score == ("C0000039", 1.0)
-        umls_entity = self.linker.umls.cui_to_entity[id_with_score[0]]
+        umls_entity = self.linker.kb.cui_to_entity[id_with_score[0]]
         assert umls_entity.concept_id == "C0000039"
         assert umls_entity.types == ["T109", "T121"]
 
@@ -61,11 +62,11 @@ class TestUmlsLinker(unittest.TestCase):
         doc.ents = (doc[2:3],)
         doc = self.linker(doc)
 
-        id_with_score = doc.ents[0]._.umls_ents[0]
+        id_with_score = doc.ents[0]._.kb_ents[0]
         assert id_with_score == ("C0000098", 1.0)
-        umls_entity = self.linker.umls.cui_to_entity[id_with_score[0]]
+        umls_entity = self.linker.kb.cui_to_entity[id_with_score[0]]
         assert umls_entity.concept_id == "C0000098"
 
     def test_linker_has_types(self):
         # Just checking that the type tree is accessible from the linker
-        assert len(self.linker.umls.semantic_type_tree.flat_nodes) == 6
+        assert len(self.linker.kb.semantic_type_tree.flat_nodes) == 6
